pyproject.toml
[project]
name = "cv_tool"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "fastapi>=0.110",
  "uvicorn[standard]>=0.27",
  "python-multipart>=0.0.9",
  "pydantic>=2.6",
  "pydantic-settings>=2.2",
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
  "python-docx>=1.1.0",
  "docxtpl>=0.16.7",
  "pymupdf>=1.23.0",
  "tenacity>=8.2.3",
  "orjson>=3.9.15"
]

[tool.uvicorn]
factory = false